refactor: move DomainErrorCode to shared lib/errors/codes so frontend can import without server deps
Extract DOMAIN_ERROR_CODES and DomainErrorCode from server/errors.ts into
apps/web/lib/errors/codes.ts. Also replaces the dynamic `code` getter on
DomainError base class with an abstract readonly property, eliminating the
static-cast and making each subclass wire up its own instance code explicitly.
