+++ apps/web/lib/trpc/testing.ts
@@ -1,6 +1,6 @@
 import { TRPCClientError } from '@trpc/client';
 
-import type { DomainErrorCode } from '@/server/errors';
+import type { DomainErrorCode } from '@/lib/errors/codes';
 import type { AppRouter } from '@/server/trpc/router';
 
 type ClientError = TRPCClientError<AppRouter>;
